python/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
websockets==12.0
loguru==0.7.2
python-dotenv==1.0.0
pipecat-ai[webrtc,daily,silero,openai,deepgram,cartesia]
